Rename health extension to smallrye-health
