--- extensions/health/deployment/src/test/java/org/jboss/shamrock/health/test/FailingUnitTest.java
@@ -1,67 +1,0 @@
-/*
- * Copyright 2018 Red Hat, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.jboss.shamrock.health.test;
-
-import static org.junit.jupiter.api.Assertions.assertEquals;
-
-import java.util.ArrayList;
-import java.util.List;
-
-import javax.enterprise.inject.Instance;
-import javax.inject.Inject;
-
-import org.eclipse.microprofile.health.Health;
-import org.eclipse.microprofile.health.HealthCheck;
-import org.eclipse.microprofile.health.HealthCheckResponse;
-import org.jboss.shamrock.test.ShamrockUnitTest;
-import org.jboss.shrinkwrap.api.ShrinkWrap;
-import org.jboss.shrinkwrap.api.asset.EmptyAsset;
-import org.jboss.shrinkwrap.api.spec.JavaArchive;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.RegisterExtension;
-
-import io.restassured.RestAssured;
-
-public class FailingUnitTest {
-
-
-    @RegisterExtension
-    static final ShamrockUnitTest config = new ShamrockUnitTest()
-            .setArchiveProducer(() ->
-                    ShrinkWrap.create(JavaArchive.class)
-                            .addClasses(FailingHealthCheck.class)
-                            .addAsManifestResource(EmptyAsset.INSTANCE, "beans.xml")
-            );
-    @Inject
-    @Health
-    Instance<HealthCheck> checks;
-
-    @Test
-    public void testHealthServlet() {
-        RestAssured.when().get("/health").then().statusCode(503);
-    }
-
-    @Test
-    public void testHealthBeans() {
-        List<HealthCheck> check = new ArrayList<>();
-        for (HealthCheck i : checks) {
-            check.add(i);
-        }
-        assertEquals(1, check.size());
-        assertEquals(HealthCheckResponse.State.DOWN, check.get(0).call().getState());
-    }
-}
