+++ extensions/smallrye-health/runtime/src/main/java/org/jboss/shamrock/smallrye/health/runtime/SmallRyeHealthServlet.java
@@ -1,0 +1,48 @@
+/*
+ * Copyright 2018 Red Hat, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.jboss.shamrock.smallrye.health.runtime;
+
+import java.io.IOException;
+
+import javax.inject.Inject;
+import javax.servlet.ServletException;
+import javax.servlet.annotation.WebServlet;
+import javax.servlet.http.HttpServlet;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+
+import io.smallrye.health.SmallRyeHealth;
+import io.smallrye.health.SmallRyeHealthReporter;
+
+
+@SuppressWarnings("serial")
+@WebServlet
+public class SmallRyeHealthServlet extends HttpServlet {
+
+    @Inject
+    SmallRyeHealthReporter reporter;
+
+    @Override
+    protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
+
+        SmallRyeHealth health = reporter.getHealth();
+        if (health.isDown()) {
+            resp.setStatus(503);
+        }
+        reporter.reportHealth(resp.getOutputStream(), health);
+    }
+}
